Pull auto approve into new analyzer
