+++ src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/tools/commandLineAnalyzer/commandLineAnalyzer.ts
@@ -3,11 +3,14 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import type { IMarkdownString } from '../../../../../../../base/common/htmlContent.js';
+import type { IDisposable } from '../../../../../../../base/common/lifecycle.js';
 import type { OperatingSystem } from '../../../../../../../base/common/platform.js';
+import type { ToolConfirmationAction } from '../../../../../chat/common/languageModelToolsService.js';
 import type { ITerminalInstance } from '../../../../../terminal/browser/terminal.js';
 import type { TreeSitterCommandParserLanguage } from '../../treeSitterCommandParser.js';
 
-export interface ICommandLineAnalyzer {
+export interface ICommandLineAnalyzer extends IDisposable {
 	analyze(options: ICommandLineAnalyzerOptions): Promise<ICommandLineAnalyzerResult>;
 }
 
@@ -17,9 +20,12 @@ export interface ICommandLineAnalyzerOptions {
 	shell: string;
 	os: OperatingSystem;
 	treeSitterLanguage: TreeSitterCommandParserLanguage;
+	terminalToolSessionId: string;
 }
 
 export interface ICommandLineAnalyzerResult {
 	readonly isAutoApproveAllowed: boolean;
-	readonly disclaimers: string[];
+	readonly disclaimers?: readonly string[];
+	readonly autoApproveInfo?: IMarkdownString;
+	readonly customActions?: ToolConfirmationAction[];
 }
